refactor(scanner): double instead of float

+++ Formal Languages and Compiler Design/FLEX scanner/example.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-    float a;
-    float b;
+    double a;
+    double b;
     a = 0xAB12;
     if (a > 10.3) {
         cout << a;
